+++ src/main/scala/org/kneelawk/imagegenerator/ImageSource.scala
@@ -1,5 +1,5 @@
 package org.kneelawk.imagegenerator
 
 trait ImageSource {
-  def apply(x: Int, y: Int): Color
+  def apply(x: Float, y: Float): Color
 }
